Data.TreeDiff.Expr

Description

This module uses Expr for richer diffs than based on Tree.

Types

data Expr

A untyped Haskell-like expression.

Having richer structure than just Tree allows to have richer diffs.

Constructors

App ConstructorName [Expr]
application

Rec ConstructorName (OMap FieldName Expr)
record constructor

Lst [Expr]
list constructor

type ConstructorName = String

Constructor name is a string

type FieldName = String

Record field name is a string too.

data EditExpr

Type used in the result of ediff.

Constructors

EditApp ConstructorName [Edit EditExpr]

EditRec ConstructorName (OMap FieldName (Edit EditExpr))

EditLst [Edit EditExpr]

EditExp Expr
unchanged tree

data Edit a

List edit operations

The Swp constructor is redundant, but it let us spot a recursion point when performing tree diffs.

Constructors

Ins a
insert

Del a
delete

Cpy a
copy unchanged

Swp a a
swap, i.e. delete + insert

exprDiff :: Expr -> Expr -> Edit EditExpr

Diff two Expr.

For examples see ediff in Data.TreeDiff.Class.